Split the main policy file into multiple files
