+++ h/auth/__init__.py
@@ -10,11 +10,11 @@ from h.auth.policy import (
     TokenAuthenticationPolicy,
 )
 from h.auth.util import default_authority
+from h.security import derive_key
 
 # We export this for the websocket to use as it's main policy
 __all__ = ("TokenAuthenticationPolicy",)
 
-from h.security import derive_key
 
 log = logging.getLogger(__name__)
