+++ tests/h/auth/policy/bearer_token_test.py
@@ -1,0 +1,50 @@
+from unittest.mock import sentinel
+
+import pytest
+
+from h.auth.policy.bearer_token import TokenAuthenticationPolicy
+from h.security import Identity
+
+
+@pytest.mark.usefixtures("user_service", "auth_token_service")
+class TestTokenAuthenticationPolicy:
+    def test_identity(self, pyramid_request, auth_token_service, user_service):
+        pyramid_request.auth_token = sentinel.auth_token
+
+        identity = TokenAuthenticationPolicy().identity(pyramid_request)
+
+        auth_token_service.validate.assert_called_once_with(sentinel.auth_token)
+        user_service.fetch.assert_called_once_with(
+            auth_token_service.validate.return_value.userid
+        )
+        assert identity == Identity(user=user_service.fetch.return_value)
+
+    def test_identity_for_webservice(self, pyramid_request, auth_token_service):
+        pyramid_request.auth_token = sentinel.decoy
+        pyramid_request.path = "/ws"
+        pyramid_request.GET["access_token"] = sentinel.access_token
+
+        TokenAuthenticationPolicy().identity(pyramid_request)
+
+        auth_token_service.validate.assert_called_once_with(sentinel.access_token)
+
+    def test_identity_returns_None_with_no_token(self, pyramid_request):
+        pyramid_request.auth_token = None
+
+        assert TokenAuthenticationPolicy().identity(pyramid_request) is None
+
+    def test_identity_returns_None_for_invalid_tokens(
+        self, pyramid_request, auth_token_service
+    ):
+        pyramid_request.auth_token = sentinel.auth_token
+        auth_token_service.validate.return_value = None
+
+        assert TokenAuthenticationPolicy().identity(pyramid_request) is None
+
+    def test_identity_returns_None_for_invalid_users(
+        self, pyramid_request, user_service
+    ):
+        pyramid_request.auth_token = sentinel.auth_token
+        user_service.fetch.return_value = None
+
+        assert TokenAuthenticationPolicy().identity(pyramid_request) is None
