+++ tests/h/auth/policy/combined_test.py
@@ -1,23 +1,16 @@
-import base64
 from unittest import mock
-from unittest.mock import sentinel
 
 import pytest
 from pyramid.interfaces import IAuthenticationPolicy
 from pyramid.security import Authenticated, Everyone
 
-from h.auth.policy import (
+from h.auth.policy._basic_http_auth import AuthClientPolicy
+from h.auth.policy.bearer_token import TokenAuthenticationPolicy
+from h.auth.policy.combined import (
     AUTH_CLIENT_API_WHITELIST,
     APIAuthenticationPolicy,
-    AuthClientPolicy,
     AuthenticationPolicy,
-    IdentityBasedPolicy,
-    RemoteUserAuthenticationPolicy,
-    TokenAuthenticationPolicy,
 )
-from h.exceptions import InvalidUserId
-from h.models.auth_client import GrantType
-from h.security import Identity
 
 API_PATHS = ("/api", "/api/foo", "/api/annotations/abc123")
 
@@ -405,282 +398,3 @@ class TestAPIAuthenticationPolicy:
         pyramid_request.matched_route.name = "api.groups"
         pyramid_request.method = "POST"
         return pyramid_request
-
-
-@pytest.mark.usefixtures("user_service")
-class TestAuthClientPolicy:
-    def test_identity(self, pyramid_request, auth_client, user_service):
-        pyramid_request.headers["X-Forwarded-User"] = sentinel.forwarded_user
-
-        identity = AuthClientPolicy().identity(pyramid_request)
-
-        user_service.fetch.assert_called_once_with(sentinel.forwarded_user)
-        assert identity == Identity(
-            auth_client=auth_client, user=user_service.fetch.return_value
-        )
-
-    def test_identify_without_forwarded_user(self, pyramid_request, auth_client):
-        pyramid_request.headers["X-Forwarded-User"] = None
-
-        identity = AuthClientPolicy().identity(pyramid_request)
-
-        assert identity == Identity(auth_client=auth_client)
-
-    def test_identity_returns_None_without_credentials(self, pyramid_request):
-        pyramid_request.headers["Authorization"] = None
-
-        assert AuthClientPolicy().identity(pyramid_request) is None
-
-    def test_identity_returns_None_if_auth_client_id_is_invalid(
-        self, pyramid_request, auth_client, db_session
-    ):
-        db_session.delete(auth_client)
-        db_session.flush()
-
-        assert AuthClientPolicy().identity(pyramid_request) is None
-
-    def test_identity_returns_None_if_auth_client_missing(
-        self, pyramid_request, auth_client, db_session
-    ):
-        db_session.delete(auth_client)
-        db_session.flush()
-
-        assert AuthClientPolicy().identity(pyramid_request) is None
-
-    def test_identify_returns_None_if_auth_client_not_secret(
-        self, auth_client, pyramid_request
-    ):
-        self.set_http_credentials(pyramid_request, "NOT A UUID", auth_client.secret)
-
-        assert AuthClientPolicy().identity(pyramid_request) is None
-
-    def test_identify_returns_None_if_auth_client_not_client_type(
-        self, auth_client, pyramid_request
-    ):
-        auth_client.grant_type = None
-
-        assert AuthClientPolicy().identity(pyramid_request) is None
-
-    def test_identify_returns_None_if_auth_client_secret_is_None(
-        self, auth_client, pyramid_request
-    ):
-        auth_client.secret = None
-
-        assert AuthClientPolicy().identity(pyramid_request) is None
-
-    def test_identify_returns_None_if_auth_client_secrets_do_not_match(
-        self, auth_client, pyramid_request
-    ):
-        self.set_http_credentials(pyramid_request, auth_client.id, "WRONG SECRET")
-
-        assert AuthClientPolicy().identity(pyramid_request) is None
-
-    def test_identify_returns_None_if_forwarded_user_is_not_found(
-        self, user_service, pyramid_request
-    ):
-        user_service.fetch.return_value = None
-
-        assert AuthClientPolicy().identity(pyramid_request) is None
-
-    def test_identify_returns_None_if_forwarded_userid_is_invalid(
-        self, user_service, pyramid_request
-    ):
-        user_service.fetch.side_effect = InvalidUserId(sentinel.invalid_id)
-
-        assert AuthClientPolicy().identity(pyramid_request) is None
-
-    def test_identify_returns_None_on_forwarded_userid_authority_mismatch(
-        self, user, auth_client, pyramid_request
-    ):
-        user.authority = "not" + auth_client.authority
-
-        assert AuthClientPolicy().identity(pyramid_request) is None
-
-    def test_unauthenticated_userid(self, pyramid_request):
-        pyramid_request.headers["X-Forwarded-User"] = "forwarded-user"
-
-        assert (
-            AuthClientPolicy().unauthenticated_userid(pyramid_request)
-            == "forwarded-user"
-        )
-
-    def test_unauthenticated_userid_returns_clientid_if_no_forwarded_user(
-        self, pyramid_request, auth_client
-    ):
-        pyramid_request.headers["X-Forwarded-User"] = None
-
-        assert (
-            AuthClientPolicy().unauthenticated_userid(pyramid_request) == auth_client.id
-        )
-
-    def test_unauthenticated_userid_returns_None_if_no_credentials(
-        self, pyramid_request
-    ):
-        pyramid_request.headers["Authorization"] = None
-        pyramid_request.headers["X-Forwarded-User"] = None
-
-        assert AuthClientPolicy().unauthenticated_userid(pyramid_request) is None
-
-    @pytest.fixture
-    def auth_client(self, factories):
-        return factories.ConfidentialAuthClient(grant_type=GrantType.client_credentials)
-
-    @pytest.fixture(autouse=True)
-    def with_credentials(self, pyramid_request, auth_client):
-        self.set_http_credentials(pyramid_request, auth_client.id, auth_client.secret)
-        pyramid_request.headers["X-Forwarded-User"] = sentinel.forwarded_user
-
-    @classmethod
-    def set_http_credentials(cls, pyramid_request, client_id, client_secret):
-        encoded = base64.standard_b64encode(
-            f"{client_id}:{client_secret}".encode("utf-8")
-        )
-        creds = encoded.decode("ascii")
-        pyramid_request.headers["Authorization"] = f"Basic {creds}"
-
-    @pytest.fixture(autouse=True)
-    def user(self, factories, auth_client, user_service):
-        user = factories.User(authority=auth_client.authority)
-        user_service.fetch.return_value = user
-        return user
-
-
-class TestIdentityBasedPolicy:
-    def test_identity_method_does_nothing(self, pyramid_request):
-        assert IdentityBasedPolicy().identity(pyramid_request) is None
-
-    @pytest.mark.parametrize(
-        "method", ("authenticated_userid", "unauthenticated_userid")
-    )
-    def test_userid_methods(self, policy, pyramid_request, identity, method):
-        assert getattr(policy, method)(pyramid_request) == identity.user.userid
-
-    @pytest.mark.parametrize(
-        "method", ("authenticated_userid", "unauthenticated_userid")
-    )
-    def test_userid_methods_return_None_if_the_identity_has_no_user(
-        self, policy, pyramid_request, identity, method
-    ):
-        identity.user = None
-
-        assert getattr(policy, method)(pyramid_request) is None
-
-    @pytest.mark.parametrize(
-        "method", ("authenticated_userid", "unauthenticated_userid")
-    )
-    def test_userid_methods_return_None_if_the_identity_is_None(
-        self, policy, pyramid_request, method
-    ):
-        policy.returned_identity = None
-
-        assert getattr(policy, method)(pyramid_request) is None
-
-    def test_effective_principals(
-        self, policy, pyramid_request, identity, principals_for_identity
-    ):
-        principals = policy.effective_principals(pyramid_request)
-
-        principals_for_identity.assert_called_once_with(identity)
-        assert principals == principals_for_identity.return_value
-
-    def test_remember_does_nothing(self, policy, pyramid_request):
-        assert policy.remember(pyramid_request, "foo") == []
-
-    def test_forget_does_nothing(self, policy, pyramid_request):
-        assert policy.forget(pyramid_request) == []
-
-    @pytest.fixture
-    def identity(self, factories):
-        return Identity(user=factories.User())
-
-    @pytest.fixture
-    def policy(self, identity):
-        class CustomPolicy(IdentityBasedPolicy):
-            returned_identity = None
-
-            def identity(self, _request):
-                return self.returned_identity
-
-        policy = CustomPolicy()
-        policy.returned_identity = identity
-
-        return policy
-
-    @pytest.fixture(autouse=True)
-    def principals_for_identity(self, patch):
-        return patch("h.auth.policy.principals_for_identity")
-
-
-@pytest.mark.usefixtures("user_service", "auth_token_service")
-class TestTokenAuthenticationPolicy:
-    def test_identity(self, pyramid_request, auth_token_service, user_service):
-        pyramid_request.auth_token = sentinel.auth_token
-
-        identity = TokenAuthenticationPolicy().identity(pyramid_request)
-
-        auth_token_service.validate.assert_called_once_with(sentinel.auth_token)
-        user_service.fetch.assert_called_once_with(
-            auth_token_service.validate.return_value.userid
-        )
-        assert identity == Identity(user=user_service.fetch.return_value)
-
-    def test_identity_for_webservice(self, pyramid_request, auth_token_service):
-        pyramid_request.auth_token = sentinel.decoy
-        pyramid_request.path = "/ws"
-        pyramid_request.GET["access_token"] = sentinel.access_token
-
-        TokenAuthenticationPolicy().identity(pyramid_request)
-
-        auth_token_service.validate.assert_called_once_with(sentinel.access_token)
-
-    def test_identity_returns_None_with_no_token(self, pyramid_request):
-        pyramid_request.auth_token = None
-
-        assert TokenAuthenticationPolicy().identity(pyramid_request) is None
-
-    def test_identity_returns_None_for_invalid_tokens(
-        self, pyramid_request, auth_token_service
-    ):
-        pyramid_request.auth_token = sentinel.auth_token
-        auth_token_service.validate.return_value = None
-
-        assert TokenAuthenticationPolicy().identity(pyramid_request) is None
-
-    def test_identity_returns_None_for_invalid_users(
-        self, pyramid_request, user_service
-    ):
-        pyramid_request.auth_token = sentinel.auth_token
-        user_service.fetch.return_value = None
-
-        assert TokenAuthenticationPolicy().identity(pyramid_request) is None
-
-
-@pytest.mark.usefixtures("user_service")
-class TestRemoteUserAuthenticationPolicy:
-    def test_unauthenticated_userid(self, pyramid_request):
-        pyramid_request.environ["HTTP_X_FORWARDED_USER"] = sentinel.forwarded_user
-
-        userid = RemoteUserAuthenticationPolicy().unauthenticated_userid(
-            pyramid_request
-        )
-
-        assert userid == sentinel.forwarded_user
-
-    def test_identity(self, pyramid_request, user_service):
-        pyramid_request.environ["HTTP_X_FORWARDED_USER"] = sentinel.forwarded_user
-
-        identity = RemoteUserAuthenticationPolicy().identity(pyramid_request)
-
-        user_service.fetch.assert_called_once_with(sentinel.forwarded_user)
-        assert identity == Identity(user=user_service.fetch.return_value)
-
-    def test_identity_returns_None_for_no_forwarded_user(self, pyramid_request):
-        pyramid_request.environ["HTTP_X_FORWARDED_USER"] = None
-
-        assert RemoteUserAuthenticationPolicy().identity(pyramid_request) is None
-
-    def test_identity_returns_None_for_no_user(self, pyramid_request, user_service):
-        pyramid_request.environ["HTTP_X_FORWARDED_USER"] = sentinel.forwarded_user
-        user_service.fetch.return_value = None
-
-        assert RemoteUserAuthenticationPolicy().identity(pyramid_request) is None
